Move state up to container
