+++ src/components/RestaurantsList.tsx
@@ -1,0 +1,3 @@
+export const RestaurantsList = () => {
+  return <div>Hello</div>;
+};
